requirements.txt
pytorch==1.12.1
cudatoolkit==11.6.0
torchvision==0.13.1
tqdm
FrEIA==0.2
torchmetrics==0.11.0
